feat: add useUserList hook for managing user list
This commit introduces a new useUserList hook, which returns an object of functions and data to handle and display the user list in the app. The hook provides access to Redux slices such as orderBy and deleteUser, to dispatch their respective actions to the application state. It also includes a function to select all the user objects stored in the redux store, which is useful for display purposes.